organized folders
Organized folders and some files
