+++ front_end/src/components/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Dashboard from '../Dashboard';
+import Dashboard from './Dashboard';
 
 
 class App extends React.Component {
